requirements.txt
mcp>=0.9.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
ollama>=0.1.6
openai>=1.12.0
psutil>=5.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
slack-sdk>=3.23.0

# Optional: For automated token extraction
# selenium>=4.15.0
